Update patch set 3
Patch Set 3:

To better assess the impact of this bug and fix, could someone from your team please help us identify the following:
Severity: Could you elaborate on the severity of this bug?  Is this a BLOCKER such that we cannot release Beryllium without it?  Is there a workaround such that we can write a release note?
Testing: Could you also elaborate on the testing of this patch?  How extensively has this patch been tested?  Is it covered by any unit tests or system tests?  
Impact: Does this fix impact any dependent projects?

Patch-set: 3
Label: Verified=0