Update patch set 19
Patch Set 19:

My objection is to the fact all of our bundles end up being dependent on karaf packaging. I suspect this will result in things like maven-shade-plugin pulling in karaf fatures -- which is wrong.

In my mind these are separate concerns: one is how the artifact is built and the other is how it is packaged (for a particular packaging system). yang-data-api depends on guava -- not on odl's packaging of guava for karaf.

Patch-set: 19