Add path string to DataValidationFailedException message
Having the path string in the message is always useful for debugging.
A separate static method is needed to work around
"call to super must be first statement in constructor".
Some messages were edited to not end in full stop.
One unit test was strengthened to assert the exact message includes the path.

Change-Id: I16dd5568808e0b323bdfcb185ba31664d493b06c
Signed-off-by: Vratko Polak <vrpolak@cisco.com>