Bug 6867: Extend yang statement parser to support different yang versions
This patch extends statement parser in order to support different versions
of yang, yin or other statement based languages. Statement parser is able
to distinguish version of each yang, yin or other source and it performs
parsing of the source in according to the specific version.
Mixed-version schemacontext (i.e. imports across versions) is supported too.

Change-Id: I06bbcbfee091a0f54480d5f5f588a9ed18a84b55
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>