Update patch set 1
Patch Set 1:

This is for investigating https://bugs.opendaylight.org/show_bug.cgi?id=7102 (will put that into commit message).
As c1 in the Bug shows, this patch made it clear that the conflict is on pcep-topology. I have not found such detail in the karaf log (linked there as [0]) without this.

Patch-set: 1