Update patch set 4
Patch Set 4:

In general, the Yangtools exceptions ideally should not rely on post-processing in Controller in order to let user know the conflicting path.

In this specific case, I agree that the LOG.warn just before org.opendaylight.controller.cluster.datastore.ShardDataTree.processNextTransaction(ShardDataTree.java:540) contains everything needed. But it is an asynchronous LOG, so the WARN appears several lines before the exception stacks, making it less easy to spot. (I missed it the first time just because the conflict can be on global-instruction-queue instead of the expected pcep-topology). Putting path to the exception message would make it pop out easier.

I think that ideally the path string should be in ConflictingModificationAppliedException message, and processNextTransaction could add Store Tx string to OptimisticLockFailedException message.
Would this Change be ok if I contributed slf4j-like pattern constructors also to Controller exceptions?

Patch-set: 4