Cleanup MandatoryLeafEnforcer and DataTreeConfiguration
This patch cleans up the implementation to simplify the NoOp
implementation, which now becomes anonymous.

Also the instantiation logic is updated to not interpret tree type
as the disabler, as that should be driven by the configuration itself.

This requires temporarily changing the configuration default to
not enforce mandatory nodes for OPERATIONAL -- that will be switched
back in a follow-up patch.

It also requires cleaning up DataTreeConfiguration so that it passes
down the correct rootPath.

Change-Id: I4e885e30433e1ae3cc43233035547d07fbd7556c
Signed-off-by: Robert Varga <rovarga@cisco.com>