Update patch set 1
Patch Set 1: Code-Review-1

This does not look right ... SchemaTracker should not be entering grouping nodes at all.

Rather than retrofitting an ill-defined method, I think we should have an explicit set of methods, which search a particular target space.

I further suspect a similar problem will ensure if we have a conflict between a container and an rpc.

I think the root problem is in how JSONNormalizedNodeStreamWriter is instantiated -- the caller (NormalizedNodeJsonBodyWriter) has already looked up the correct SchemaNode, yet we lose that lookup and are re-doing it again just based on SchemaPath -- not knowing what we are really looking for.

Patch-set: 1
Label: Code-Review=-1