Update patch set 17
Patch Set 17:

(4 comments)

"Furthermore on the SemVer thing ... neither 1.2 nor 2.0 of YANG is on the NETMOD map, so I think we'd be better off with a simple two-entry enum. 2.0 is likely to be a completely different beast, and 1.2 is not scoped (so could end up being a surprise)." 

But then whole statement parser would be limited to this two yang versions only. I think current solution is more scalable, not tied to yang only and I think, it costs nothing more than proposed enum. Furthermore we have exposed CustomStatementParserBuilder and maybe someone could make for example its own extension and support more versions etc...

Patch-set: 17
Label: Verified=0