Update patch set 22
Patch Set 22: Code-Review-1

It is scalable, but honestly ... given the amount of changes we are doing for 1.1, is it really probable that it would be useful for 1.2 or 2.0?

I honestly do not think 2.0 will have a compatible metamodel, which really means model.api will have to be redone from scratch. That would imply SemVer.major  would not see any use.

Furthermore, and this is critical, nothing in 6020 or 7950 language suggests anything about semantic versioning, so we MUST NOT assume that version number has SemVer semantics -- it certainly does not have the correct string format.

Let us just define a YangVersion enum in yang-common with VERSION_1 and VERSION_1_1, both of which should have a getReference() returning "RFC6020" and "RFC7950" respectively. There should also be a parse(String) static factory method throwing java.text.ParseException, understanding "1" and "1.1".

That will also allow us to use (the very fast) EnumMap for any version-specific mappings.

Patch-set: 22
Label: Code-Review=-1