BUG-6652: generate features.xml with karaf-maven-plugin
This is a simplified version of
https://git.opendaylight.org/gerrit/38638 without any changes to
bundles; it just delegates the actual features.xml generation to the
plugin. This requires a modified version of Karaf to reduce the
resulting dependencies
(https://github.com/skitt/karaf/tree/remove-duplicate-bundles). The
new features are generated for Karaf 4 only. (The existing
features.xml is preserved for Karaf 3.)

Change-Id: I4cc4692459f637e06e64a1a2e432448f57c32e4c
Signed-off-by: Stephen Kitt <skitt@redhat.com>