Update patch set 5
Patch Set 5:

Right, but the idea here is that we need to identify the source (passed in by the user) to the user.

Hence all we really need to do is have StatementStreamSource extend Identifiable<SourceIdentifier> and use that. Then the exception will pick up the identifier from StatementStreamSource, which is always available and we're done.

YinStatementStreamSource already has this. YinStatementStreamSourceImpl is just a proxy on top of it. YangStatementStreamSourceImpl needs to be fixed, but for that we need to weed out all the weird public constructors.

Patch-set: 5