Eliminate superfluous onFullDefinitionDeclared overrides
This patch moves invocation of the validator back into AbstractStatementSupport,
rendering majority of overrides useless, as all they did was call (then no-op)
super and then the validator.

This also fixes problems if a subclass redefines the validator to be non-existent
and fails to override (again) onFullDefinitionDeclared().

Drive-by cleanup reduces verbosity of specification implementation classes by
importing the nested interface directly.

The end effect is that it is again possible to reason about hooks being performed
in statements by looking at what implementation overrides there are.

Change-Id: Ibed4cd5eeb128e69637fcf124331aeeeaeb7b4ea
Signed-off-by: Robert Varga <rovarga@cisco.com>