Reuse substatements across phases
Now that we can address substatements accurately, we can reuse
the same statements instead of throwing them out and re-instantiating
them.

We still need to switch the root if the YangVersion changes, as the
statement support objects need to be reemitted.

FIXME: this patch does not retain order of declared statements, which
       is a problem coming from the fact we keep them in different
       lists.

Change-Id: I6fd3e2e2de41305457958673b77a72073215940e
Signed-off-by: Robert Varga <rovarga@cisco.com>