Hide INodeBase.mainnode
This field is accessed either as CAS or READ, with writes being
a falso positiive, in that they are performed only at initialization
time.

Refactor access patterns to hide the updater and the field, passing
down explicit initializer from the subclass.

Also remove INodeBase.prev(), as it is not called from anywhere.

Change-Id: I3bfcc7e7a606b3a0622ed04b99abbc67e41eb407
Signed-off-by: Robert Varga <rovarga@cisco.com>