BUG-7464: Fix inconsistent identity check on removal
Removal of values through Map.remove(Object, Object) was
inconsistent in that the SNode case uses v.equals() whereas
the LNode case uses identity check. This is inconsistent
and runs contrary to the interface specitication, which calls
for using Objects.equals() -- which boils down to v.equals()
just as the SNode case does.

Fix the LNode case to also invoke v.equals() to conform to
specification.

Change-Id: I37bead74b52ad4a1c74d18099dc705afad67665d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>