Update patch set 9
Patch Set 9:

I do not consider this solution as ideal, but I have tried several approaches how to solve this issue and only this one seems as feasible. 
If you consider this solution as not good enough, I try to return to it some time later. Any ideas how to make it better are welcome.

Patch-set: 9
Label: Verified=0