Update patch set 7
Patch Set 7:

I think, we can also fix Yang 1.0 in the same way as Yang 1.1 and I think it will be really better, if the behavior is the same for the both Yang versions. However, I only wanted to raise a discussion about this, since RFC6020 is not clear about this. I will also write to NETMOD mailing list, but what is your preference (@Vatko, @Robert) ?

Patch-set: 7
Label: Verified=0