Update patch set 8
Patch Set 8: Code-Review-1

Omitting 'case' statement is pure syntactic sugar and models performing an augment should be written as if the case statement were present.

According to RFC6020 section 10, it is fair to add new nodes, which may require expanding an implicit case into an explicit one. This action should not break downstream models.

This is the same situation as RPC's input/output statements, I do believe this should be handled the same way in both YANG 1.0 and 1.1.

Patch-set: 8
Label: Code-Review=-1