Update patch set 9
Patch Set 9:

> Omitting 'case' statement is pure syntactic sugar and models
 > performing an augment should be written as if the case statement
 > were present.
 > 
 > According to RFC6020 section 10, it is fair to add new nodes, which
 > may require expanding an implicit case into an explicit one. This
 > action should not break downstream models.
 > 
 > This is the same situation as RPC's input/output statements, I do
 > believe this should be handled the same way in both YANG 1.0 and
 > 1.1.

Yes, this is what Martin Bjorklund said. I am posting his answer once again: "Yes, this works the same way in both YANG 1 and 1.1.  The new sentence was added as a clarification to 7950."

So, I changed this patch to work for both Yang 1 and 1.1 in the same way.

Patch-set: 9
Label: Verified=0