(WIP) Bug 8126 - Some augments are not being processed
If a node is marked as mandatory (e.g. "leaf" with "mandatory true;"), yang statement parser
does not perform augmentation of such node into a target node even though one of ancestors of
the target node is a non-mandatory choice or a non-mandatory list from the same namespace as
the mandatory node and so in fact the node should not be considered as mandatory any more and
augmentation should be performed.

We also increase severity of the augment error log from debug to error.

Change-Id: Icb8554ba61ce0aa0fa25541dd5675f3aa8ba3d1b
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>