(WIP) Bug 7037 - Improve mapping of YANG extensions
Each extension statement definition is able to create an unknown statement
form of a regular yang statement. Created definitions of regular yang
statements are created only one per each yang statement and extension.
So we do not create a new ModelDefinedStatementDefinition and a corresponding
UnknownStatementImpl.Definition for each instance of yang statement when
it is placed in an unknown statement's body.

- need to be tested ...

Change-Id: I5613eb36273e6b512637fb14456c64b023df3f1f
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>