Bug 7037 - Improve mapping of YANG extensions
Each statement definition of regular yang statement is able to create
an unknown statement definition of itself and only once pre each yang
statement and only in case if it is necessary. So we do not create
a new ModelDefinedStatementDefinition and a corresponding
UnknownStatementImpl.Definition for each instance of yang statement when
it is placed in an unknown statement's body.

Change-Id: I5613eb36273e6b512637fb14456c64b023df3f1f
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>