Update patch set 1
Patch Set 1: Code-Review-1

I think this needs to be the other way around: statement definitions need to be hosted in the parent extension, so that:

some:ext1 {
    container aaa;
}

other:ext2 {
    container aaa;
}

Results in two distinct definitions for container aaa. Note that for example tailf-common defines a meta-language which can express which statements are valid inside a definition.

Patch-set: 1
Label: Code-Review=-1