Update patch set 1
Patch Set 1: Code-Review+1

Maybe it would be slightly better to introduce new method (or at least new boolean argument) for lookups when we want to perform lookup regardless of  'case' nodes (i.e. when 'case' node is not present in supplied SchemaPath) so that the original method will not be affected. However, the impact is minimal and only for really special cases (that I think nobody is using the method in such way), so I vote +1.

Patch-set: 1
Reviewer: Gerrit User 2745 <2745@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1