Bug 8675: Fix design flaws of the new XML parser
New XML parser assumes that YANG-modeled data in the parsed XML source
are wrapped in a YANG-independent root element which is not correct.
Actually, the root element of such XML document should correspond
to a specific YANG data node defined in the provided YANG model.

Another issue with the parser is the SchemaNode parameter that is used when
creating an XmlParserStream instance. Currently, it represents a schema
node that is the parent node of the root XML element. However, this
parameter should correspond to the root XML element itself.

Fix these issues.

Adjust existing unit tests accordingly.

Change-Id: I900717f652cd93e36c229bc91d3854b5ab99f09a
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>