Bug 8675: Fix a design flaw of the new XML parser
The actual flaw is that the parser wraps the parentNode
SchemaNode inside a CompositeNodeDataWithSchema which emits
only its child nodes into the provided NormalizedNodeStreamWriter.
The parentNode itself is not emitted which is a problem.
Therefore we need to check the parentNode type and put it
into a corresponding ...NodeDataWithSchema.

Bug description on bugzilla has been updated with a more
thorough explanation.

Adjust existing unit tests accordingly.

Add javadoc to ...NodeDataWithSchema classes.

Change-Id: I900717f652cd93e36c229bc91d3854b5ab99f09a
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>