Update patch set 2
Patch Set 2:

> This looks like we are changing behavior -- since this is already
 > used by RESTCONF, that worries me.
 > 
 > Isn't it perhaps so that the codec is used in two different ways?

Actually, the RESTCONF project is not using the new XML parser yet. The changes in this patch were made with the purpose of making the new XML parser closer to the old one in terms of its usage and thus to make the migration of RESTCONF/NETCONF easier. The only problem is the BGPCEP project which has already been using the XmlParserStream in its ConfigLoaderImpl class. The changes in this patch cause breakage in there, but I believe that it can be fixed. I just need to discuss this issue with BGPCEP devs first.

Patch-set: 2
Label: Verified=0