(WIP) Bug 2444 - Yin export of all YANG statements
Currently yang-export exports effective world of yang into yin,
what is not correct, because lot of information is not accesible
in effective world anymore.

Short list of some issues:
- we are unable to export QName/SchemaPath with original prefixes in some cases
- we are unable to export if-feature statements
- currenly, we are unable to export submodule and includes statements correctly
- we are unable to get original form of a statement in some cases (e.g. applied
  deviations etc.)

So the correct solution is to export declared world of yang,
where all information is accessible and in raw form and that is
exactly what we need.

Change-Id: Ic2bc559623d9e72c59e2ed7f6a797cd67933ada6
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>